[llvm] add pseudo instruction and change activation record allocation
for LMUL > 1, the stack need allocate the multiple of vlenb,
for now, I just specify the vectorsize by vscale vector's
getKnownMinSize() / 8 * vlenb, there are maybe bugs that I
haven't discovered. And I optimize the assembly code that
In the previous implementation, before the instruction use the
frameindex, there are mv a0, a0, which is redundancy. This mv
instruction was introduced by the addi instruciton, which the
compiler optimize the addi to mv.