change register class's diagnostic string
The tablegen backend for MC asm parse is a little terrible,
for the different register class that has the same register set,
the tablegen will generate same MCK enumeration value, which will
make the MCK_FPR32 disappear, and get MCK_FPR16, because the
tablegen backend use a map to record the registerclass info. For the
same register set, it will map to the same value. When print the
MCK enumeration value, the tablegen only print one of it. So I add
a record diagnositic string to make the tablegen generate MCK_FPR32.