fix error when load non-vector stack object
the error is that when load non-vector stack object, the compiler use the sp
as base address. But, when the stack has vector object allocated, the sp
is not the real sp, all the stack object must use the fp as base address.
in the getFrameIndexReference, I decided if using sp or fp by if the stack object
is vector object. So, when the object is not the vector object, the base address register
is sp. This caused the error.